Sat-Func Initializers: Use RawTableEndPoints for KRORG
This commit reimplements the findKrorg helper function in terms of
the RawTableEndPoints aggregate.  In particular, this commit aims to
underscore the fact that KRORG is the relative permeability of oil
in the Oil/Gas system at the displacing saturation

    1 - Sgcr - Swco

This fact is hopefully more apparent in this revised than the
existing implementation since we now rely explicitly on evaluating
the 'KrogColumn' at an index derived from looking up that saturation
value in the pertinent saturation column.