Sat-Func Initializers: Use RawTableEndPoints for KRORW
This commit reimplements the findKrorw helper function in terms of
the RawTableEndPoints aggregate.  In particular, this commit aims to
underscore the fact that KRORW is the relative permeability of oil
in the Oil/Water system at the displacing saturation

    1 - Swcr - Sgco

This fact is hopefully more apparent in this revised than the
existing implementation since we now rely explicitly on evaluating
the 'KrowColumn' at an index derived from looking up that saturation
value in the pertinent saturation column.