Sat-Func Initializers: Use RawTableEndPoints for KRWR
This commit reimplements the findKrwr helper function in terms of
the RawTableEndPoints aggregate.  In particular, this commit aims to
underscore the fact that KRWR is the relative permeability of water
at the displacing saturation

    1 - Socr - Sgco

This fact is hopefully more apparent in this revised than the
existing implementation since we now rely explicitly on evaluating
the 'KrwColumn' at an index derived from looking up that saturation
value in the pertinent saturation column.