Sat-Func Initializers: Use RawTableEndPoints for KRGR
This commit reimplements the findKrgr helper function in terms of
the RawTableEndPoints aggregate.  In particular, this commit aims to
underscore the fact that KRGR is the relative permeability of gas at
the displacing saturation

    1 - Socr - Swco

This fact is hopefully more apparent in this revised than the
existing implementation since we now rely explicitly on evaluating
the 'KrgColumn' at an index derived from looking up that saturation
value in the pertinent saturation column.