SOGCR: Pull SGOF Handling Into Initializer
This is in any case needed to calculate correct KRGR default values
in runs using family I saturation function keywords (S*OF) and also
removes the need to apply the special 'subtract_swl' operation in
this case.  Overall this is a simplification of the code since we
now have all saturation end-points available when initializing
saturation function field properties.