Restart: Prepare to Output Well/Group Guide Rates
This commit identifies vector items in XWEL and XGRP that hold guide
rate values at the well and group levels, and adds functionality to
output those values to the restart file.  We do not identify such
value at the FIELD level.  As with other summary-like quantities, we
expect to pick up the fully converted values from a SummaryState
object.

Calculating and storing the guide rate quantities in SummaryState is
the subject of future work.