Add Restart I/O Support for Well's THP
This commit adds support for saving and restoring a well's THP
quantitiy.  We put the THP into the summary facility's list of
"required restart quantities" and pick up the value converted to
output units from the SummaryState.

On restart we restore the THP value to internal (SI) units and
transport the value to the simulator in data::Well::thp.  While
here, also add the THP value to the RstWell object although
currently not strictly needed in that context.