Group Data: Conditionally Enable Debug Output
This commit switches the group control debug output to std::cout
into a conditional mode that is controlled by a file-wide PP symbol

    ENABLE_GCNTL_DEBUG_OUTPUT

that we set to zero here.  This allows the developers to turn on
additional debug output when needed without confusing those who are
not actively working on the feature.