Output: Transport Guiderate Values to Output Layer
This commit introduces datamembers of type

    Opm::data::GuideRateValue

to the Opm::data::Well and Opm::data::GroupData structures.
Collectively, these datamembers enable transporting guiderate values
calculated elsewhere--e.g., in the opm-simulators module--to the
output layer for final processing and output to the summary and/or
the restart files.

We support outputting both production and injection guiderate values
at the group level.  Guiderate values at the well level are
interpreted as production or injection according to the well type.