Don't Write RSM File for Substeps
If the last report step happens to be split into multiple substeps
(e.g., convergence problems or TUNING settings &c), then we will
write the RSM file for each substep which will slow down the
process.  This commit explicitly ensures that we don't output the
RSM file unless it is the last substep of the final report step.