PVTO: Switch to Using {fmt} Backend for Formatting
This commit switches the custom diagnostic message formatting based
on std::ostringstream to using a formatter implemented in terms of
the fmt::format function.  This removes the custom indentation from
earlier.  Note that we output at most four records to the console,
including a message limit notification if applicable, and that this
limit is currently not adjustable by the user.  We do on the other
hand output all flipped records to the .PRT file.

New Example Diagnostic Output:

  * To the Console

    Warning: Non-Monotonic Oil Formation Volume Factor Detected in Keyword PVTO, PVTNUM=1
    In /path/to/PVTO.INCL line 6
    Record  9: FVF 1.234 at RS 123.456 is not greater than FVF 2.345 at RS 121.212
    Record 10: FVF 1.233 at RS 123.457 is not greater than FVF 1.234 at RS 123.456
    Record 11: FVF 1.233 at RS 123.460 is not greater than FVF 1.233 at RS 123.457
    Report limit reached, see PRT-file for additional records.

  * In the PRT File (Unlimited Records)

    Warning: Non-Monotonic Oil Formation Volume Factor Detected in Keyword PVTO, PVTNUM=1
    In /path/to/PVTO.INCL line 6
    Record  9: FVF 1.234 at RS 123.456 is not greater than FVF 2.345 at RS 121.212
    Record 10: FVF 1.233 at RS 123.457 is not greater than FVF 1.234 at RS 123.456
    Record 11: FVF 1.233 at RS 123.460 is not greater than FVF 1.233 at RS 123.457
    Record 12: FVF 1.233 at RS 123.461 is not greater than FVF 1.233 at RS 123.460
    Record 13: FVF 1.233 at RS 123.470 is not greater than FVF 1.233 at RS 123.461

The length of each 'Record' line is 78 characters as of this change
(assuming METRIC conventions, FIELD conventions not tested).  The
length of the 'In ...' line depends on the number of characters in
the full filepath.