Don't Overrun Formatting Buffer in 'doub_string_ecl'
We would write outside the allocated buffer, and fail to write the
nul terminator, when presented with inputs of very small size
(denormalized numbers in the order of 1.0e-101 or smaller).  Switch
to using std::snprintf and expand the buffer size to work around
this problem.

While here, also fix a subtle issue in determining whether or not we
should insert the exponent character 'D' into the formatted output.
Since we add one upon formatting the number, we also need to use a
non-symmetric condition for the 'D' character lest we fail to output
correctly formatted versions of

    -2.9440373045442E-100 (=> -0.29440373045442D-99)
    -2.9440373045442E+99  (=> -0.29440373045442+100)