Configure Aquifer Summary Output Needed for Restart
The summary vectors

    AAQP (pressure), AAQR (rate), AAQT (volume)

must be available for output to the restart file's "XAAQ" array for
all analytic aquifers.  Additionally, the summary vectors

    AAQTD (dimensionless time), AAQPD (dimensionless pressure)

must be available for restart output to "XAAQ" in the case of
Carter-Tracy aquifers.

This commit ensures that those values are always calculated and
stored in the SummaryState.