Install cmake package configuration to correct location with lib.
If we ship a library than we are architecture dependent and should
allow for installations for multiple architectures. For this to work
the package configuration files need to be installed in an
architecture aware location (e.g. lib/arch/cmake/module_name on
Debian). Just like the pkg-config files. This is done with this patch.