Provide Means of Linearising All Aquifer IDs in a Model
This commit adds two helper functions

    std::vector<int> analyticAquiferIDs(const AquiferConfig&)
    std::vector<int> numericAquiferIDs(const AquiferConfig&)

which return the numeric IDs of the model's analytic and numeric
aquifers respectively.  These are useful for configuring aquifer
summary keywords of the form

    AAQR
    /

that do not list individual aquifer IDs and for identifying issues
in the input file when a summary keyword that does list its aquifer
IDs refers to an aquifer of incorrect category (e.g., ID of numeric
aquifer in an analytic aquifer summary keyword).

While here, also add more unit test coverage to the AquiferConfig
object.  In particular, add a unit test that includes aquifers at
all.