Track UDA Active Status for Group Injection Controls
This commit adds a new member function, updateUDQActive, to the
group injection properties.  This is inspired by the function of the
same name in the group production properties and fills the same
purpose.  This, in turn, enables tracking whether a GCONINJE keyword
activates a user-defined argument.  That status is in turn needed to
generate the 'IUAD' restart vector in the restart file in the face
of models for which the only UDA is in GCONINJE.