Account for RSEG Size Depending on Number of Active Phases
The number of data items per well segment in the RSEG restart vector
turns out to depend on the number of active phases (Oil, Gas, Water)
in the simulation run.  In particular, the following relations hold

  * Single phase (Oil, Gas, or Water):          NRSEGZ = 126
  * Two phases (Oil/Gas, Oil/Water, Gas/Water): NRSEGZ = 134
  * Three phases (Oil/Gas/Water):               NRSEGZ = 146

This commit implements that variability.

The output layer does not currently use any items above index 110 so
this does not alter the contents of RSEG, only the size, for 1p or
2p simulation runs.