Check ALQ Type When Accumulating Gas Lift Injection Rate
This commit adds a guard to the xGLIR accumulation procedure.  In
particular, we do not accumulate anything unless the ALQ type is gas
rate.  We do however output a well-level value if the ALQ type is
gas/liquid ratio.  This guard requires passing the Schedule object
as part of the 'fn_args', so update the call sites accordingly.