Output WGLIR=0 Unless ALQ=GRAT
This reduces the risk of users misinterpreting the curves since the
unit of measure when plotting the vector will unconditionally be
surface volume rate for gas.

We will revisit this decision if we gain the ability to capture the
actual lift gas injection rate independent of ALQ type.

Suggested by: [at]tskille