Add 'Inplace' Slot for Dynamic Pore Volume
The 'PoreVolume' value is supposed to be calculated at reference
conditions (i.e., no rock compressibility or temperature effects),
but being able to report the non-reference condition value is
beneficial.  This commit adds a slot for accumulating that value per
FIP region and transporting the aggregated values to the PRT output
system.