Use Previous Output Time to Decide Next Output Time
This commit switches the criterion for BASIC=4 and BASIC=5 to using
the difference between the previous and the next output time instead
of "just" the next output time.  The original criteria

    year % freq == 0
    month % freq == 0

don't do what we want.  BASIC=4 and BASIC=5 should output every
'freq' year (=4) or 'freq' month (=5) of simulated time.