Add Means of Pruning Deactivated Aquifer Connections
This commit adds operation pruneDeactivatedAquiferConnections at
the EclipseState and AquiferConfig levels.  This is intended as a
stop-gap solution to a sequencing problem that arises from MINPV
processing.  In general we must not form connections to analytic
aquifers until the set of active grid cells is fully established.

Pruning cells that become deactivated as part of MINPV handling
nevertheless covers the common case of connection boxes applying
only to the outer shell.  Solving the problem in general will
however require delaying forming the analytic aquifer connections
until after the active cells are fully known.