Add Predicate for Potential Rates of W/G Name
This commit adds a new predicate, GuideRate::hasPotentials(), that
checks whether or not rates exist in the 'potentials' array for a
particular well or group name.  If neither explicit guide rate
values nor potential rates exist for a particular "wgname", then it
is an error to call GuideRate::get() and that member function will
throw an exception.  This new predicate provides a way of avoiding
that exception.