Provide External Access to Well-Order Sorting
This commit extracts the 'NameOrder' sorting operation to a new
public member function, WellMatcher::sort(vector<string>).  The
immediate use-case for this function is to provide a stable sorting
order for wells that are matched in an ACTIONX block.  Well IDs
often end up in various output arrays in the restart files--e.g., in
the IUAP array the matching wells are used in a WCON* keyword that
employs a user-defined argument as part of setting its targets or
limits--and we should strive to keep those arrays as deterministic
as possible.