Load Field-Wide Constraints From Restart File
If constraints are applied at the field level, e.g., through the
GCONINJE keyword, then we must incorporate those constraints into
FIELD when restarting a simulation run.  This commit takes a step in
this direction by explicitly loading the *GRP array entries
pertaining to FIELD and special casing

    Schedule::addGroup(RstGroup, time)

to look for "rst_group.name == FIELD".  In that case, we update the
injection and production properties of the existing FIELD group as
needed.