Don't Access an Empty Optional Object
The maximum supply of lift gas and/or the maximum total gas flow
rate of a group may be defaulted in the context of lift gas
optimisation.  In that case, the std::optional<> objects will be
empty and we must not access the contained object through the
value() member function.  We output the sentinel value "-10" in this
case.