Don't Access Well BHP Reference Depth If Unavailable
This commit ensures that we do not access the well's BHP reference
depth unless that value is well defined.  If the input deck does not
define an explicit BHP reference depth (WELSPECS(5)), and if the
well does not have any active reservoir connections, then we are
supposed to output the usual sentinel value (-1.0E+20) as the datum
depth in 'SWEL'.

Account for this altered semantics in the restart reader too.