Do not repeat warning for every COMPDAT record.
Without this, for every COMPDAT record that refer to an inactive cell
there will be two warnings, one for the cell and one that the well
still has no connections and will remain SHUT (provided the well did
not have any in the first place). With this, the second warning (for
the well) will be printed once, after the complete COMPDAT keyword
has been processed.